requirements.txt
# Python: 3.11

numpy
pandas
matplotlib
seaborn
scipy
scikit-learn
jupyterlab
ipykernel

automata-lib==8.4.0
cached-method==0.1.0
click==8.1.8
contourpy==1.3.0
cvxopt==1.3.2
cycler==0.12.1
deprecation==2.1.0
fonttools==4.54.1
frozendict==2.4.6
graphviz==0.20.1
intervaltree==3.1.0
joblib==1.4.2
kiwisolver==1.4.7
lxml==5.3.0
markdown-it-py==3.0.0
mdurl==0.1.2
memory-profiler==0.61.0
networkx==3.4.2
packaging==24.1
pillow==11.0.0
pm4py==2.7.11.14
psutil==6.1.0
pydotplus==2.0.2
pyfiglet==1.0.2
pyparsing==3.2.0
pytz==2024.2
rich==14.0.0
shellingham==1.5.4
six==1.16.0
sortedcontainers==2.4.0
style==1.1.0
threadpoolctl==3.5.0
tqdm==4.66.6
typer==0.15.2
tzdata==2024.2
update==0.0.1